+++ src/modules.js
@@ -159,7 +159,7 @@ var LibraryManager = {
 
     if (USE_WEBGPU) {
       libraries.push('library_webgpu.js');
-      libraries.push('library_emscripten_webgpu.js');
+      libraries.push('library_html5_webgpu.js');
     }
 
     if (BOOTSTRAPPING_STRUCT_INFO) {
